requirements.txt
Requests==2.31.0
streamlit==1.29.0
llama_cpp_python==0.2.24